Add files via upload
GEANT4 simulations for the experiment at Occidental College in March 2023.   It contains the geometry and materials for a GEM and a Cf-252 source.  It was built from the GdScintV3 package.

This package updates GdScintV3 by creating a sensitive detector and a text output file to list the particles that deposit energy in the sensitive detector.